% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{agrupar_tecnologias_barras}
\alias{agrupar_tecnologias_barras}
\title{Agrupar tabela com muitas categorias}
\usage{
agrupar_tecnologias_barras(tab, other_level = "Demais tecnologias")
}
\arguments{
\item{tab}{Uma tibble saída de uma função calcula_eixo_**}

\item{other_level}{String que será passada na função [forcats::fct_lump_n()]}
}
\description{
Recebe uma tabela saída de calcula_eixo_** e reduz o número de
valores distintos da coluna x.
}
