% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcular_eixo_y_serie}
\alias{calcular_eixo_y_serie}
\title{Cria a tabela base para qualquer grafico de séries de tempo}
\usage{
calcular_eixo_y_serie(
  tab,
  eixo_y,
  var_cor = NULL,
  nome_var_cor = NULL,
  other_level = "Demais tecnologias"
)
}
\arguments{
\item{tab}{Tabela retornada por uma chama da função \code{\link[dplyr::tbl]{tbl}}.
Essa função aplicará \code{\link[dplyr::collect]{collect}} neste objeto.}

\item{eixo_y}{String descrevendo o que entrará no eixo y do gráfico. coluna "y" da tabela de saída.
Pode assumir os valores "valor", "porcPIB", "quantidade", "procProj". Esse parâmetro define os valores
das colunas ylab, e format_func. Veja a seção "Value".}

\item{var_cor}{String descrevendo o que dará a cor de cada linhas. Coluna "grupo" da tabela de saída.
Deve ser um do nome das colunas de [tab].}
}
\value{
Uma tibble contendo as colunas ano, y, grupo, ylab, y_format, format_func. Em função do parâmetro
eixo_y, [calcular_eixo_y_barras_emp] retorna uma label (ylab) e o valor formatado da coluna y (format_func)
com base no que estiver preenchido na tabela [tab_vis_dispendios],
criada em data-raw/tab_auxiliares.R.
}
\description{
Cria a tabela base para qualquer grafico de séries de tempo. O resultado 
é uma tibble pronta para ser usada em alguma função gráfica e a entrada é o resultado
de uma chamada da funcao \code{\link[dplyr::tbl]{tbl}}.
}
