% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graficos.R
\name{e_tooltip}
\alias{e_tooltip}
\title{Tooltip do echarts4r}
\usage{
e_tooltip(e, trigger = c("item", "axis"), ...)
}
\arguments{
\item{e}{An echarts4r object as returned by e_charts or a proxy as returned by echarts4rProxy.}

\item{trigger}{What triggers the tooltip, one of item or item.}

\item{...}{Any other option to pass.}
}
\value{

}
\description{
A função e_tooltip do echarts4r não permite que se passe os
parametros adicionais diretamente nos ... Essa função modifica o comportamento
da função original para que seja possivel fazer e_tooltip(e, "axis", list(...))
}
