% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graficos.R
\name{ec_heatmap}
\alias{ec_heatmap}
\title{Constrói um heatmap em echarts4r}
\usage{
ec_heatmap(tab, xlab)
}
\arguments{
\item{tab}{Uma tibble com colunas x, y e v. x e y representam os valores dos
eixos x e y, respectivamente. v representa a cor do heatmap}
}
\value{
Um objeto do tipo echarts4r plotando y contra x, pintado pela coluna v
}
\description{
Constrói um heatmap em echarts4r
}
