% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graficos.R
\name{ec_hist}
\alias{ec_hist}
\title{Constrói um gráfico de área em echarts4r}
\usage{
ec_hist(tab)
}
\arguments{
\item{tab}{Uma tibble com colunas x e y. Caso essa coluna esteja agrupada, os gráficos
de área serão construídos respeito esse agrupamento.}
}
\value{
Um objeto do tipo echarts4r plotando y contra x, respeitando eventuais agrupamentos
do parâmetro tab.
}
\description{
Constrói um gráfico de área em echarts4r
}
