% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graficos.R
\name{escolher_fomatador}
\alias{escolher_fomatador}
\title{A partir de uma tabela pronta para colocar num grafico, retorna o formatador
correspondente}
\usage{
escolher_fomatador(tab, JS = TRUE)
}
\arguments{
\item{tab}{Uma tibble produzida por uma função calcular_eixo_**}

\item{JS}{Variavel logica. O retorno da funcao deve ser um código
em JavaScript ou o nome de uma função declarada em "www/script.js"?}
}
\value{
String contendo um código JS para ser executado posteriormente
}
\description{
A partir de uma tabela pronta para colocar num grafico, retorna o formatador
correspondente
}
