% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatar_dinheiro}
\alias{formatar_dinheiro}
\title{Função formatadora de valores monetários}
\usage{
formatar_dinheiro(x)
}
\arguments{
\item{x}{Vetor do tipo numérico}
}
\description{
Wrapper da função [scales::dollar] que padroniza dados
para o formato brasileiro
}
