% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatar_numero}
\alias{formatar_numero}
\title{Função formatadora de números > 1}
\usage{
formatar_numero(x, acc = 1)
}
\arguments{
\item{x}{Vetor do tipo numérico}
}
\description{
Wrapper da função [scales::number] que padroniza números
para o formato brasileiro
}
