% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formatar_porc}
\alias{formatar_porc}
\title{Função formatadora de porcentagem}
\usage{
formatar_porc(x)
}
\arguments{
\item{x}{Vetor do tipo numérico}
}
\description{
Wrapper da função [scales::percent] que padroniza dados em %
para o formato brasileiro
}
