% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parametros}
\alias{parametros}
\title{Parametros gerais do app}
\usage{
parametros()
}
\description{
Essa função concentra todos os parametros que afetam várias 
partes do app. Atualmente estão mapeadas (a) altura dos containers dos gráficos,
que é passada nas chamadas da função echarts4rOutput (b) tamanho da fonte das
legendas e (c) a altura do grid passado em algumas chamadas da funcao e_grid
}
