% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graficos.R
\name{plot_arvore}
\alias{plot_arvore}
\title{Constrói um treemap em echarts4r}
\usage{
plot_arvore(tab, height = "95\%", periodo = "")
}
\arguments{
\item{tab}{Tibble contendo as colunas name, children e value}

\item{height}{Altura do gráfico. Parâmetro passado para a função e_grid}
}
\value{
Um objeto do tipo echarts4r plotando y contra x, pintado pela coluna grupo
}
\description{
Constrói um treemap em echarts4r
}
