% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graficos.R
\name{plot_dist_barras_emp}
\alias{plot_dist_barras_emp}
\title{Constrói um gráfico de barras empilhadas em echarts4r}
\usage{
plot_dist_barras_emp(tab, height = "80\%")
}
\arguments{
\item{tab}{Tibble saída da função calcular_eixo_y_barras_emp}
}
\value{
Um objeto do tipo echarts4r plotando y contra x, pintado pela coluna grupo
}
\description{
Constrói um gráfico de barras empilhadas em echarts4r
}
