% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graficos.R
\name{plot_rosca}
\alias{plot_rosca}
\title{Constrói um gráfico de rosca em echarts4r}
\usage{
plot_rosca(tab)
}
\arguments{
\item{tab}{Tibble saída da função calcular_eixo_y_barras}
}
\value{
Um objeto do tipo echarts4r plotando y contra x, pintado pela coluna grupo
}
\description{
Constrói um gráfico de rosca em echarts4r
}
