% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graficos.R
\name{plot_sankey}
\alias{plot_sankey}
\title{Constrói um sankey plot em echarts4r}
\usage{
plot_sankey(
  tab,
  height = "95\%",
  ano = "",
  animation = TRUE,
  definir_cores = FALSE
)
}
\arguments{
\item{tab}{Um tibbl produzida pela função calcular_eixo_y_barras_emp}

\item{height}{Altura do gráfico. Parâmetro passado para a função e_grid}

\item{ano}{String contendo o ano que será exibido na tooltip}

\item{animation}{Booleana. O gráfico executará uma animação ao ser construído?}

\item{definir_cores}{Booleana. O gráfico deve usar o vetor cores_epe()}
}
\value{
Um objeto do tipo echarts4r plotando y contra x, pintado pela coluna grupo
}
\description{
Constrói um sankey plot em echarts4r
}
