% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_graficos.R
\name{plot_serie_linhas}
\alias{plot_serie_linhas}
\title{Constrói um série de linhas do tempo em echarts4r}
\usage{
plot_serie_linhas(tab)
}
\arguments{
\item{tab}{Uma tabela saída de calcular_eixo_y_serie}
}
\value{
Um objeto do tipo echarts4r plotando y contra ano, pintado pela coluna grupo
}
\description{
Constrói um série de linhas do tempo em echarts4r
}
