% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tabelas.R
\name{tab_projetos_min}
\alias{tab_projetos_min}
\title{Parâmetros de estilo da tabela de detalhamentos do explorador de projetos}
\usage{
tab_projetos_min(tab)
}
\description{
Parâmetros de estilo da tabela de detalhamentos do explorador de projetos
}
