% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tabelas.R
\name{tabela_dispendios}
\alias{tabela_dispendios}
\title{Parâmetros de estilo da tabela de dispendios ao longo dos anos}
\usage{
tabela_dispendios(tab)
}
\description{
Parâmetros de estilo da tabela de dispendios ao longo dos anos
}
