% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tabelas.R
\name{tabela_dispendios_agrup}
\alias{tabela_dispendios_agrup}
\title{Parâmetros de estilo da tabela de dispendios agrupados ao longo dos anos}
\usage{
tabela_dispendios_agrup(tab, nome_x = "Tecnologia")
}
\description{
Parâmetros de estilo da tabela de dispendios agrupados ao longo dos anos
}
