% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tabelas.R
\name{tabela_projetos}
\alias{tabela_projetos}
\title{Parâmetros de estilo da tabela de projetos}
\usage{
tabela_projetos(tab, col_def = NULL, colunas, onclick, pagination = FALSE)
}
\description{
Parâmetros de estilo da tabela de projetos
}
